engine: type the lane-gauge side as unsigned int
ApplyLaneGaugeValueAndBackground took its side as unsigned long long, so
every use of it needed a cast (a mask on the record index, a narrowing for
the gauge call, and another for the side test). The disassembly only ever
reads the argument as a 32-bit register, so it is an unsigned int side
index: retype it, drop all three casts, and name the second-player side
constant.

+++ Project/ScoreTracker.h
@@ -92,7 +92,7 @@ public:
      * @param uSide The player side.
      * @ghidraAddress 0x149324
      */
-    void ApplyLaneGaugeValueAndBackground(float flValue, unsigned long long uSide);
+    void ApplyLaneGaugeValueAndBackground(float flValue, unsigned int uSide);
 
     /**
      * @brief Arms a side's score-digit roll-up on the given player-field layer.
